Fix ConVar and ConCommands

+++ public/tier1/iconvar.h
@@ -112,11 +112,11 @@ public:
 
 	// Accessors.. not as efficient as using GetState()/GetInfo()
 	// if you call these methods multiple times on the same IConVar
-	virtual bool IsFlagSet( int nFlag ) const = 0;
+	virtual bool IsFlagSet( int64 nFlag ) const = 0;
 
 	virtual int GetSplitScreenPlayerSlot() const = 0;
 	
-	virtual void AddFlags( int flags ) = 0;
+	virtual void AddFlags( int64 flags ) = 0;
 };
